Restructure Code
New Structure:
cmd/*        - Command Definitions
pkg/config   - configuration
pkg/download - dowload logic
pkg/extract  - extract (tar) data post download
pkg/optname  - constants for option names (arguments)
pkg/version  - version information

main.go - Entry Point with basic "Run the program" code
