+++ cmd/version.go
@@ -1,15 +1,18 @@
-package version
+package cmd
 
 import (
 	"fmt"
+
 	"github.com/spf13/cobra"
+
+	"github.com/replicate/pget/pkg/version"
 )
 
-var CMDVersion *cobra.Command = &cobra.Command{
+var VersionCMD = &cobra.Command{
 	Use:   "version",
 	Short: "Print the version information",
 	Long:  "Print the version information",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Printf("pget Version %s - Build Time %s\n", GetVersion(), BuildTime)
+		fmt.Printf("pget Version %s - Build Time %s\n", version.GetVersion(), version.BuildTime)
 	},
 }
