+++ main.go
@@ -2,54 +2,24 @@ package main
 
 import (
 	"fmt"
-
 	"os"
 	"path/filepath"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/replicate/pget/config"
-	"github.com/replicate/pget/download"
-	"github.com/replicate/pget/extract"
-	"github.com/replicate/pget/optname"
-	"github.com/replicate/pget/version"
+	"github.com/replicate/pget/cmd"
+	"github.com/replicate/pget/pkg/config"
+	"github.com/replicate/pget/pkg/download"
+	"github.com/replicate/pget/pkg/extract"
+	"github.com/replicate/pget/pkg/optname"
 )
 
-const longDesc = `
-pget
-
-PGet is a high performance, concurrent file downloader built in Go. It is designed to speed up and optimize file
-downloads from cloud storage services such as Amazon S3 and Google Cloud Storage.
-
-The primary advantage of PGet is its ability to download files in parallel using multiple threads. By dividing the file
-into chunks and downloading multiple chunks simultaneously, PGet significantly reduces the total download time for large
-files.
-
-If the downloaded file is a tar archive, PGet can automatically extract the contents of the archive in memory, thus
-removing the need for an additional extraction step.
-
-The efficiency of PGet's tar extraction lies in its approach to handling data. Instead of writing the downloaded tar
-file to disk and then reading it back into memory for extraction, PGet conducts the extraction directly from the
-in-memory download buffer. This method avoids unnecessary memory copies and disk I/O, leading to an increase in
-performance, especially when dealing with large tar files. This makes PGet not just a parallel downloader, but also an
-efficient file extractor, providing a streamlined solution for fetching and unpacking files.
-`
-
 func main() {
-	cmd := &cobra.Command{
-		Use:   "pget [flags] <url> <dest>",
-		Short: "pget",
-		Long:  longDesc,
-		RunE:  mainFunc,
-		PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-			return config.PersistentStartupProcessFlags()
-		},
-		Args: cobra.ExactArgs(2),
-	}
-	config.AddFlags(cmd)
-	cmd.AddCommand(version.CMDVersion)
-	if err := cmd.Execute(); err != nil {
+	config.AddFlags(cmd.RootCMD)
+	cmd.RootCMD.RunE = mainFunc
+	cmd.RootCMD.AddCommand(cmd.VersionCMD)
+	if err := cmd.RootCMD.Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
